+++ apps/fluent-tester/src/E2E/Button/pages/ButtonPageObject.ts
@@ -1,18 +1,24 @@
 import {
   BUTTON_TESTPAGE,
-  BUTTON_TEST_COMPONENT,
-  BUTTON_NO_A11Y_LABEL_COMPONENT,
+  BUTTON_TEST_COMPONENT_DEPRECATED,
+  BUTTON_NO_A11Y_LABEL_COMPONENT_DEPRECATED,
   HOMEPAGE_BUTTON_BUTTON,
-  BUTTON_ON_PRESS,
+  BUTTON_ON_PRESS_DEPRECATED,
 } from '../../../FluentTester/TestComponents/Button/consts';
 import { BasePage, By } from '../../common/BasePage.win';
+
+/* This enum gives the spec file an EASY way to interact with SPECIFIC UI elements on the page.
+ * The spec file should import this enum and use it when wanting to interact with different elements on the page. */
+export const enum ButtonSelector {
+  PrimaryButton, //this._primaryComponent
+}
 
 class ButtonPageObject extends BasePage {
   /******************************************************************/
   /**************** UI Element Interaction Methods ******************/
   /******************************************************************/
   didOnClickCallbackFire(): boolean {
-    const callbackText = By(BUTTON_ON_PRESS);
+    const callbackText = By(BUTTON_ON_PRESS_DEPRECATED);
     browser.waitUntil(
       () => {
         return callbackText.isDisplayed();
@@ -27,6 +33,19 @@ class ButtonPageObject extends BasePage {
     return callbackText.isDisplayed();
   }
 
+  /* Sends a Keyboarding command on a specific UI element */
+  sendKey(buttonSelector: ButtonSelector, key: string): void {
+    this.getButtonSelector(buttonSelector).addValue(key);
+  }
+
+  /* Returns the correct WebDriverIO element from the Button Selector */
+  getButtonSelector(buttonSelector?: ButtonSelector): WebdriverIO.Element {
+    if (buttonSelector == ButtonSelector.PrimaryButton) {
+      return this._primaryComponent;
+    }
+    return this._primaryComponent;
+  }
+
   /*****************************************/
   /**************** Getters ****************/
   /*****************************************/
@@ -39,11 +58,11 @@ class ButtonPageObject extends BasePage {
   }
 
   get _primaryComponent() {
-    return By(BUTTON_TEST_COMPONENT);
+    return By(BUTTON_TEST_COMPONENT_DEPRECATED);
   }
 
   get _secondaryComponent() {
-    return By(BUTTON_NO_A11Y_LABEL_COMPONENT);
+    return By(BUTTON_NO_A11Y_LABEL_COMPONENT_DEPRECATED);
   }
 
   get _pageButton() {