+++ packages/theming/apple-theme/src/appleColors.macos.ts
@@ -4,6 +4,7 @@ import { AppleSemanticPalette, FluentUIApplePalette } from './appleColors.types.
 import { PlatformColor, DynamicColorMacOS, ColorWithSystemEffectMacOS } from 'react-native-macos';
 import { Appearance } from 'react-native';
 import { createMacOSAliasTokens } from './createMacOSAliasTokens';
+import { getIsHighContrast } from './appleHighContrast.macos';
 
 /** Creates a Palette of PlatformColors defined for macOS */
 export function getAppleSemanticPalette(): AppleSemanticPalette {
@@ -58,7 +59,7 @@ export function getAppleSemanticPalette(): AppleSemanticPalette {
 function getFluentUIApplePalette(): FluentUIApplePalette {
   const appearance = Appearance.getColorScheme();
   const mode = getCurrentAppearance(appearance, 'light');
-  const macOSAliasColorTokens = createMacOSAliasTokens(mode);
+  const macOSAliasColorTokens = createMacOSAliasTokens(mode, getIsHighContrast());
 
   return {
     blue10: '#4F6BED',
@@ -256,16 +257,10 @@ function getFluentUIApplePalette(): FluentUIApplePalette {
       dark: '#291F07',
     }),
 
-    brandForegroundDisabled: DynamicColorMacOS({
-      light: '#2525253F',
-      dark: '#FFFFFF3F',
-    }),
-    brandBackgroundDisabled: DynamicColorMacOS({
-      light: '#2525253F',
-      dark: '#5656567F',
-    }),
-
     brandedBackground: macOSAliasColorTokens.brandBackground,
+    neutralBackgroundDisabled: macOSAliasColorTokens.neutralBackgroundDisabled,
+    brandForegroundDisabled: macOSAliasColorTokens.brandForeground1Disabled,
+    brandForegroundPressed: macOSAliasColorTokens.brandForeground1Pressed,
     brandBackgroundPressed: macOSAliasColorTokens.brandBackgroundPressed,
     brandBackgroundHovered: macOSAliasColorTokens.brandBackgroundHover,
     neutralBackgroundInverted: macOSAliasColorTokens.neutralBackgroundInverted,
@@ -278,9 +273,12 @@ function getFluentUIApplePalette(): FluentUIApplePalette {
     neutralBackground3: macOSAliasColorTokens.neutralBackground3,
     neutralStrokeDisabled: macOSAliasColorTokens.neutralStrokeDisabled,
     transparentBackground: macOSAliasColorTokens.transparentBackground,
+    transparentStroke: macOSAliasColorTokens.transparentStroke,
+    neutralForegroundDisabled: macOSAliasColorTokens.neutralForegroundDisabled,
+    neutralForeground2Pressed: macOSAliasColorTokens.neutralBackground2Pressed,
+    neutralBackground2Pressed: macOSAliasColorTokens.neutralBackground2Pressed,
   };
 }
-
 /** Creates a palette of colors for the apple theme, given the FluentUI Apple Palette and Apple Semantic Palette
  * The fallback palette is loaded while we wait for  the native theming module to load, or if the module is not found
  */
@@ -309,7 +307,6 @@ export function fallbackApplePalette(): ThemeColorDefinition {
     variantBorder: applePlatform.separatorColor,
     variantBorderHovered: applePlatform.separatorColor,
     defaultStateBackground: applePlatform.controlBackgroundColor,
-
     errorText: fluentUIApple.dangerPrimary,
     warningText: fluentUIApple.warningPrimary,
     errorBackground: fluentUIApple.dangerTint10,
@@ -337,7 +334,7 @@ export function fallbackApplePalette(): ThemeColorDefinition {
     buttonBackgroundCheckedHovered: fluentUIApple.neutralBackground3,
     buttonBackgroundPressed: ColorWithSystemEffectMacOS(fluentUIApple.neutralBackground3, 'pressed'),
     buttonBackgroundDisabled: ColorWithSystemEffectMacOS(fluentUIApple.neutralBackground3, 'disabled'),
-    buttonBorder: fluentUIApple.transparentBackground,
+    buttonBorder: fluentUIApple.transparentStroke,
     buttonText: fluentUIApple.neutralForeground3,
     buttonTextHovered: fluentUIApple.neutralForeground3,
     buttonTextChecked: fluentUIApple.neutralForeground3,
@@ -350,13 +347,13 @@ export function fallbackApplePalette(): ThemeColorDefinition {
 
     primaryButtonBackground: fluentUIApple.communicationBlue,
     primaryButtonBackgroundHovered: fluentUIApple.communicationBlue,
-    primaryButtonBackgroundPressed: ColorWithSystemEffectMacOS(fluentUIApple.communicationBlue, 'pressed'),
-    primaryButtonBackgroundDisabled: fluentUIApple.brandBackgroundDisabled,
+    primaryButtonBackgroundPressed: fluentUIApple.brandBackgroundPressed,
+    primaryButtonBackgroundDisabled: fluentUIApple.neutralBackgroundDisabled,
     primaryButtonBorder: 'transparent',
     primaryButtonBorderFocused: 'transparent',
     primaryButtonText: fluentUIApple.neutralForegroundInverted,
     primaryButtonTextHovered: fluentUIApple.neutralForegroundInverted,
-    primaryButtonTextPressed: ColorWithSystemEffectMacOS(fluentUIApple.neutralForegroundInverted, 'pressed'),
+    primaryButtonTextPressed: fluentUIApple.neutralForegroundInverted,
     primaryButtonTextDisabled: fluentUIApple.brandForegroundDisabled,
 
     accentButtonBackground: fluentUIApple.communicationBlue,
@@ -389,33 +386,33 @@ export function fallbackApplePalette(): ThemeColorDefinition {
     /* ControlColorTokens */
 
     // Set the default button tokens to match the Acrylic Button style
-    defaultBackground: fluentUIApple.neutralBackground3,
-    defaultBorder: 'transparent',
-    defaultContent: fluentUIApple.neutralForeground3,
+    defaultBackground: fluentUIApple.neutralBackground2,
+    defaultBorder: fluentUIApple.neutralStroke2,
+    defaultContent: fluentUIApple.neutralForeground2,
     defaultIcon: fluentUIApple.neutralForeground3, //GH:728 Icon doesn't support PlatformColor
 
     defaultHoveredBackground: fluentUIApple.neutralBackground3,
-    defaultHoveredBorder: 'transparent',
+    defaultHoveredBorder: fluentUIApple.neutralStroke2,
     defaultHoveredContent: fluentUIApple.neutralForeground3,
     defaultHoveredIcon: fluentUIApple.neutralForeground3, //GH:728 Icon doesn't support PlatformColor
 
     defaultFocusedBackground: fluentUIApple.neutralBackground3,
-    defaultFocusedBorder: 'transparent',
+    defaultFocusedBorder: fluentUIApple.neutralStroke2,
     defaultFocusedContent: fluentUIApple.neutralForeground3,
     defaultFocusedIcon: fluentUIApple.neutralForeground3, //GH:728 Icon doesn't support PlatformColor
 
-    defaultPressedBackground: ColorWithSystemEffectMacOS(fluentUIApple.neutralBackground3, 'pressed'),
-    defaultPressedBorder: 'transparent',
-    defaultPressedContent: ColorWithSystemEffectMacOS(fluentUIApple.neutralForeground3, 'pressed'),
+    defaultPressedBackground: fluentUIApple.neutralBackground2Pressed,
+    defaultPressedBorder: fluentUIApple.neutralStroke2,
+    defaultPressedContent: fluentUIApple.neutralForeground2,
     defaultPressedIcon: fluentUIApple.neutralForeground3, //GH:728 Icon doesn't support PlatformColor
 
-    defaultDisabledBackground: fluentUIApple.neutralStrokeDisabled,
-    defaultDisabledBorder: 'transparent',
-    defaultDisabledContent: ColorWithSystemEffectMacOS(fluentUIApple.neutralForeground3, 'pressed'),
+    defaultDisabledBackground: fluentUIApple.neutralBackground2,
+    defaultDisabledBorder: fluentUIApple.neutralStroke2,
+    defaultDisabledContent: fluentUIApple.neutralForegroundDisabled,
     defaultDisabledIcon: fluentUIApple.neutralForeground3, //GH:728 Icon doesn't support PlatformColor
 
     ghostBackground: 'transparent',
-    ghostBorder: 'transparent',
+    ghostBorder: fluentUIApple.transparentStroke,
     ghostContent: fluentUIApple.communicationBlue,
     ghostIcon: fluentUIApple.communicationBlue,
 
@@ -431,7 +428,7 @@ export function fallbackApplePalette(): ThemeColorDefinition {
 
     ghostPressedBackground: 'transparent',
     ghostPressedBorder: 'transparent',
-    ghostPressedContent: ColorWithSystemEffectMacOS(fluentUIApple.communicationBlue, 'deepPressed'),
+    ghostPressedContent: fluentUIApple.brandForegroundPressed,
     ghostPressedIcon: fluentUIApple.neutralForeground3, //GH:728 Icon doesn't support PlatformColor
 
     ghostDisabledBackground: 'transparent',
@@ -440,7 +437,7 @@ export function fallbackApplePalette(): ThemeColorDefinition {
     ghostDisabledIcon: fluentUIApple.brandForegroundDisabled,
 
     brandedBackground: fluentUIApple.brandedBackground,
-    brandedBorder: 'transparent',
+    brandedBorder: fluentUIApple.transparentStroke,
     brandedContent: fluentUIApple.neutralForegroundInverted,
     brandedIcon: fluentUIApple.neutralForegroundInverted,
 
@@ -459,7 +456,7 @@ export function fallbackApplePalette(): ThemeColorDefinition {
     brandedPressedContent: fluentUIApple.neutralForegroundInverted,
     brandedPressedIcon: fluentUIApple.neutralForegroundInverted,
 
-    brandedDisabledBackground: fluentUIApple.brandBackgroundDisabled,
+    brandedDisabledBackground: fluentUIApple.neutralBackgroundDisabled,
     brandedDisabledBorder: 'transparent',
     brandedDisabledContent: fluentUIApple.brandForegroundDisabled,
     brandedDisabledIcon: fluentUIApple.neutralForeground3, //GH:728 Icon doesn't support PlatformColor
@@ -495,7 +492,7 @@ export function fallbackApplePalette(): ThemeColorDefinition {
     defaultPressedSecondaryContent: fluentUIApple.neutralForegroundInverted,
 
     checkboxBackground: fluentUIApple.communicationBlue,
-    checkboxBackgroundDisabled: fluentUIApple.brandBackgroundDisabled,
+    checkboxBackgroundDisabled: fluentUIApple.neutralBackgroundDisabled,
     checkboxBorderColor: fluentUIApple.gray600,
     checkmarkColor: fluentUIApple.neutralForegroundInverted,