+++ packages/theming/apple-theme/src/appleHighContrast.macos.ts
@@ -1,0 +1,9 @@
+let isHighContrastEnabled = false;
+
+export function setIsHighContrast(isHighContrast: boolean) {
+  isHighContrastEnabled = isHighContrast;
+}
+
+export function getIsHighContrast() {
+  return isHighContrastEnabled;
+}