+++ packages/components/Button/CHANGELOG.md
@@ -1,8 +1,38 @@
 # Change Log - @fluentui-react-native/button
 
-This log was last generated on Wed, 26 Jan 2022 20:31:04 GMT and should not be manually modified.
+This log was last generated on Tue, 15 Feb 2022 22:48:31 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.22.0
+
+Tue, 15 Feb 2022 22:48:31 GMT
+
+### Minor changes
+
+- Move experimental files to component folder (ruaraki@microsoft.com)
+
+## 0.21.20
+
+Wed, 09 Feb 2022 20:14:50 GMT
+
+### Patches
+
+- Move file (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/interactive-hooks to v0.13.0
+- Bump @fluentui-react-native/pressable to v0.8.19
+
+## 0.21.19
+
+Wed, 02 Feb 2022 02:29:07 GMT
+
+### Patches
+
+- Bump @uifabricshared/foundation-compose to v1.11.9
+- Bump @fluentui-react-native/icon to v0.11.6
+- Bump @fluentui-react-native/interactive-hooks to v0.12.2
+- Bump @fluentui-react-native/pressable to v0.8.18
+- Bump @fluentui-react-native/text to v0.11.9
 
 ## 0.21.18