+++ packages/theming/win32-theme/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui-react-native/win32-theme
 
-This log was last generated on Fri, 14 Jan 2022 21:49:07 GMT and should not be manually modified.
+This log was last generated on Wed, 02 Feb 2022 02:29:07 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.14.10
+
+Wed, 02 Feb 2022 02:29:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.9.9
+- Bump @fluentui-react-native/theme-tokens to v0.12.0
+- Bump @fluentui-react-native/theming-utils to v0.8.9
 
 ## 0.14.9